LuatTN update function
