+++ pharmacy/src/main/java/com/c1221g1/pharmacy/dto/import_invoice/SupplierDto.java
@@ -2,6 +2,7 @@ package com.c1221g1.pharmacy.dto.import_invoice;
 
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Pattern;
 
 public class SupplierDto {
     private String supplierId;
@@ -13,15 +14,17 @@ public class SupplierDto {
 
     @NotNull
     @NotEmpty
+    @Pattern(regexp = "^(0?)(3[2-9]|5[6|8|9]|7[0|6-9]|8[0-6|8|9]|9[0-4|6-9])[0-9]{7}$")
     private String supplierPhone;
 
     @NotNull
     @NotEmpty
+    @Pattern(regexp = "^[\\w-\\.]+@([\\w-]+\\.)+[\\w-]{2,4}$")
     private String supplierEmail;
 
     private String note;
     private boolean flag;
-//    private List<ImportInvoice> importInvoiceList;
+
 
     public SupplierDto() {
     }
