+++ pharmacy/src/test/java/com/c1221g1/pharmacy/controller/controllerSupplier/SupplierRestController_deleteSupplier.java
@@ -1,4 +1,4 @@
-package com.c1221g1.pharmacy.controller;
+package com.c1221g1.pharmacy.controller.controllerSupplier;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
